% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{puny_decode}
\alias{puny_decode}
\title{Decode Punycode domains to Unicode with normalization}
\usage{
puny_decode(domains)
}
\arguments{
\item{domains}{character vector of Punycode domains}
}
\value{
a character vector of Unicode domains
}
\description{
Decode Punycode domains to Unicode with normalization
}
