% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{puny_encode}
\alias{puny_encode}
\title{Encode Unicode domains to Punycode}
\usage{
puny_encode(domains)
}
\arguments{
\item{domains}{character vector of Unicode domains}
}
\value{
a character vector of Punycode domains
}
\description{
Encode Unicode domains to Punycode
}
